+++ rust/main/agents/relayer/src/msg/op_batch.rs
@@ -355,10 +355,16 @@ mod tests {
                     ..Default::default()
                 },
                 consider_null_transaction_receipt: false,
+                grpc_urls: None,
+                solidity_grpc_urls: None,
+                energy_multiplier: None,
             }),
             metrics_conf: Default::default(),
             index: Default::default(),
+            confirmations: Default::default(),
+            chain_id: Default::default(),
             ignore_reorg_reports: false,
+            native_token: Default::default(),
         };
 
         // https://explorer.hyperlane.xyz/message/0x29160a18c6e27c2f14ebe021207ac3f90664507b9c5aacffd802b2afcc15788a