+++ rust/main/chains/hyperlane-tron/src/lib.rs
@@ -5,14 +5,21 @@
 #![deny(clippy::unwrap_used, clippy::panic)]
 #![deny(clippy::arithmetic_side_effects)]
 
-use ethers_signers::LocalWallet;
-
 pub use {
-    config::ConnectionConf, contracts::TronInterchainGasPaymaster, contracts::TronMailbox,
-    contracts::TronMailboxIndexer, contracts::TronMerkleTreeHook,
-    contracts::TronMerkleTreeHookIndexer, contracts::TronValidatorAnnounce,
-    ism::TronAggregationIsm, ism::TronInterchainSecurityModule, ism::TronMultisigIsm,
-    ism::TronRoutingIsm, provider::TronProvider, provider::TronProviderForLander,
+    config::ConnectionConf,
+    contracts::TronInterchainGasPaymaster,
+    contracts::TronMailbox,
+    contracts::TronMailboxIndexer,
+    contracts::TronMerkleTreeHook,
+    contracts::TronMerkleTreeHookIndexer,
+    contracts::TronValidatorAnnounce,
+    ism::TronAggregationIsm,
+    ism::TronInterchainSecurityModule,
+    ism::TronMultisigIsm,
+    ism::TronRoutingIsm,
+    provider::TronProvider,
+    provider::TronProviderForLander,
+    signer::{TronSigners, TronSignersError},
 };
 
 mod config;
@@ -20,13 +27,13 @@ mod contracts;
 mod error;
 mod ism;
 mod provider;
+mod signer;
 mod utils;
 
 #[allow(clippy::unwrap_used)]
 mod interfaces;
 
 /// The signer type used for Tron chain interactions
-/// This is an alias for `LocalWallet` from the `ethers_signers` crate
-pub type TronSigner = LocalWallet;
+pub type TronSigner = TronSigners;
 
 pub(crate) use {config::*, contracts::*, error::*, provider::*, utils::*};