+++ typescript/http-registry-server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @hyperlane-xyz/http-registry-server
 
+## 25.3.2
+
+### Patch Changes
+
+- @hyperlane-xyz/sdk@25.3.2
+- @hyperlane-xyz/utils@25.3.2
+
 ## 25.3.1
 
 ### Patch Changes